[HUDI-5538] Fix ContinuousFileSource and ITTestDataStreamWrite for flink 1.16 support
